+++ R/freqMap.R
@@ -21,8 +21,13 @@
 #'   overall contribution to the total concentration. Note that for options
 #'   other than \dQuote{frequency}, it is necessary to also provide the name of
 #'   a pollutant. See function [openair::cutData()] for further details.
-#' @param latitude The decimal latitude.
-#' @param longitude The decimal longitude.
+#' @param latitude The decimal latitude. If not provided, latitude will be
+#'   automatically inferred from data by looking for a column named \dQuote{lat}
+#'   or \dQuote{latitude} (case-insensitively).
+#' @param longitude The decimal longitude. If not provided, longitude will be
+#'   automatically inferred from data by looking for a column named
+#'   \dQuote{lon}, \dQuote{lng}, \dQuote{long}, or \dQuote{longitude}
+#'   (case-insensitively).
 #' @param provider The base map(s) to be used. See
 #'   \url{http://leaflet-extras.github.io/leaflet-providers/preview/} for a list
 #'   of all base maps that can be used. If multiple base maps are provided, they
@@ -55,8 +60,8 @@
 freqMap <- function(data,
                     pollutant = "nox",
                     statistic = "mean",
-                    latitude = "lat",
-                    longitude = "lon",
+                    latitude = NULL,
+                    longitude = NULL,
                     provider = "OpenStreetMap",
                     type = "default",
                     cols = "jet",
@@ -68,6 +73,14 @@ freqMap <- function(data,
                     fig.height = 4,
                     ...) {
   . <- NULL
+
+  latlon <- assume_latlon(
+    data = data,
+    latitude = latitude,
+    longitude = longitude
+  )
+  latitude <- latlon$latitude
+  longitude <- latlon$longitude
 
   data <-
     prepMapData(