+++ R/percentileMap.R
@@ -1,15 +1,20 @@
 #' Percentile roses on interactive leaflet maps
 #'
 #' @param data A data frame. The data frame must contain the data to plot a
-#'   [openair::percentileRose()], which includes wind speed (\code{ws}), wind direction
-#'   (\code{wd}), and the column representing the
-#'   concentration of a pollutant. In addition, \code{data} must include a
-#'   decimal latitude and longitude.
+#'   [openair::percentileRose()], which includes wind speed (\code{ws}), wind
+#'   direction (\code{wd}), and the column representing the concentration of a
+#'   pollutant. In addition, \code{data} must include a decimal latitude and
+#'   longitude.
 #' @param pollutant The column name(s) of the pollutant(s) to plot. If multiple
 #'   pollutants are specified, they can be toggled between using a "layer
 #'   control" interface.
-#' @param latitude The decimal latitude.
-#' @param longitude The decimal longitude.
+#' @param latitude The decimal latitude. If not provided, latitude will be
+#'   automatically inferred from data by looking for a column named \dQuote{lat}
+#'   or \dQuote{latitude} (case-insensitively).
+#' @param longitude The decimal longitude. If not provided, longitude will be
+#'   automatically inferred from data by looking for a column named
+#'   \dQuote{lon}, \dQuote{lng}, \dQuote{long}, or \dQuote{longitude}
+#'   (case-insensitively).
 #' @param provider The base map(s) to be used. See
 #'   \url{http://leaflet-extras.github.io/leaflet-providers/preview/} for a list
 #'   of all base maps that can be used. If multiple base maps are provided, they
@@ -40,8 +45,8 @@
 #' }
 percentileMap <- function(data,
                           pollutant = "nox",
-                          latitude = "lat",
-                          longitude = "lon",
+                          latitude = NULL,
+                          longitude = NULL,
                           provider = "OpenStreetMap",
                           type = "default",
                           cols = "jet",
@@ -53,6 +58,14 @@ percentileMap <- function(data,
                           fig.height = 4,
                           ...) {
   . <- NULL
+
+  latlon <- assume_latlon(
+    data = data,
+    latitude = latitude,
+    longitude = longitude
+  )
+  latitude <- latlon$latitude
+  longitude <- latlon$longitude
 
   data <-
     prepMapData(